requirements.txt
flake8~=7.1.1
google-generativeai~=0.8.2
jsonlines~=4.0.0
python-dotenv~=1.0.1
Sphinx~=8.0.2
tqdm~=4.66.5
